Added coarse grid source and receiver interpolation to rays.cpp

+++ include/rays.h
@@ -150,7 +150,7 @@ public:
 
     // Insert source functions
     int insertSource(std::shared_ptr<Data3D<T>> source, bool maptype); ///< Insert source position
-    int insertSource(std::shared_ptr<Data3D<T>> source, bool maptype, int traceno); ///< Insert source position for a particular trace number
+    int insertSource(std::shared_ptr<Data3D<T>> source, bool maptype, size_t traceno); ///< Insert source position for a particular trace number
     void insertResiduals(std::shared_ptr<Data3D<T>> source, bool maptype); ///< Insert traveltime residuals
     void insertImageresiduals(T *res); ///< Insert image residuals
     void clearTT(); ///< Reset traveltime array
